removed react table and toggle grid/list only in tweets
