+++ frontend/src/features/tweets/components/Tweets.tsx
@@ -1,12 +1,14 @@
 import { Tweet } from '../types';
 import { useTweets } from '../api/getTweets';
 import SingleTweet from './SingleTweet';
+import ToggleListGrid from '../../dashboard/components/toggleListGrid';
+import React, { useState } from 'react';
 
-interface Props {
-	isList: { activeLayout: string };
-}
+const Tweets = () => {
+	const [isList, setListActive] = useState({
+		activeLayout: 'list',
+	});
 
-const Tweets = ({ isList }: Props) => {
 	const { isLoading, error, data: tweets } = useTweets();
 
 	if (isLoading) {
@@ -17,13 +19,31 @@ const Tweets = ({ isList }: Props) => {
 		return <div>An error occurred: {error.message}</div>;
 	}
 
+	const handleClick = (e: React.MouseEvent<HTMLButtonElement>, name: string) => {
+		e.preventDefault();
+		if (isList.activeLayout !== name) {
+			setListActive({
+				activeLayout: name,
+			});
+		}
+	};
+
 	const displayTweets = () => {
 		return (
-			<div className={isList.activeLayout === 'list' ? 'grid gap-4' : 'grid md:grid-cols-2 lg:grid-cols-3 gap-6'}>
-				{tweets.map((tweet: Tweet) => (
-					<SingleTweet key={tweet.id} tweet={tweet} />
-				))}
-			</div>
+			<>
+				<div className="pb-3 flex justify-end">
+					<ToggleListGrid isList={isList} handleClick={handleClick} />
+				</div>
+				<div
+					className={
+						isList.activeLayout === 'list' ? 'grid gap-4' : 'grid md:grid-cols-2 lg:grid-cols-3 gap-6'
+					}
+				>
+					{tweets.map((tweet: Tweet) => (
+						<SingleTweet key={tweet.id} tweet={tweet} />
+					))}
+				</div>
+			</>
 		);
 	};
